Test default configurations
Now that we can easily overwrite config settings with -E, we can use the
default configs in the packaging tests. This has the advantage that we test the
default configuration files.

This PR does this for Linux and Darwin. For Windows we need a slightly
different approach, because the tests start the beats as services, so we don't
control the flags.