Zypper ignore gpg sig (Suse) (#64)
* Update to use sha512

Starting with RC1, we publish sha512 files instead of sha1 files. This
updates beats-tester for that change.

* Ignore GPG in zypper (suse)

This seems to be required to get it working on staged builds, which
contain signatures. Better would be to actually import the GPG sig, but
we currently don't do that for other platforms either.