Use test-beat role for all Beats
Update Metricbeat, Heartbeat, Filebeat, and Winlogbeat to use the test-beat roles.

Update run-settings for staging to 6.2.0 BC3.

Copy logs/config/output to host even on error. This uses a task block with an "always" condition (like a try/finally) in order to ensure that log files, output files, and config are copied to the localhost.

Add workaround for https://github.com/elastic/beats/issues/6214 on Darwin.