Use /stats to fetch metrics instead of logs
Instead of using the "Total non-zero metrics" log lines use the HTTP /stats endpoint to fetch metrics. The main benefit is that zero value metrics are not filtered. This allows the tests to assert that the metrics actually exist.

A side-effect of not using logs is that we lose a check that the beat truly shut down. We should add another assertion to check for "<beatname> stopped." in the logs.