Fix open handle metric tests
This is a follow up to #92 that added test for the open handle metric. This fixes a few issues with the tests. The linux and windows assertions assumed that the common assertions were run first but this isn't the case because the files are mutually exclusive (the most specific one found us used).