Override BEAT_LOG_OPTS in systemd unit file
This overrides the defaults in the unit file for out services so that we get a log file. The tests currently assert that the log exists.

We could extend the tests to assert that logs are written to journald in the future.

Relates to https://github.com/elastic/beats/pull/8942