Keystore testing from package
Create a keys in the keystore and start the beat with a configuration
that use the key, if the key is missing the beat will refuse to start.
This assert that the path used by the packages is correct for the CLI
and the started service.

This is to stop regression like in https://github.com/elastic/beats/pull/11497