* Adapt output file name to latest beats changes.
This fixes the tests for beats when checking for the output files.
Relates to changes from https://github.com/elastic/beats/pull/28927/
Add output ndjson suffix to APM Server output (unclear why this is necessary).

* Increase timeout for output file and ensure it is written.

The output file is only written when some actual values are processed. Enable
self-instrumentation for APM to ensure the output file is written and increase
timeout. For beats, enable metrics logging explicitly.

* Disable failing beats issues

A follow up is necessary to fix and enable these tests if still useful.